feat: migrate createStudyHandler, requestResearchHandler, askHandler to TypeScript
Phase 4 Stage 4 (1/3): first 3 handler migrations.
- Added handler context types (SlashCommandContext, BlockActionContext, ViewSubmissionContext)
- Typed LLM helpers, block builders, and form extraction
- Converted function expressions to declarations
