--- backend/src/helpers/slack/commands/requestResearchHandler.js
@@ -1,368 +1,0 @@
-const { requestResearchModal } = require("../ui/requestResearchModal");
-const { createStudyModal } = require("../ui/createStudyModal");
-const { createStudyFromRequestModal } = require("../ui/createStudyFromRequestModal");
-const { researchBriefModal } = require("../ui/researchBriefModal");
-const { getConfigRepo, YAML_TEMPLATE_PATH, fetchFileFromRepo, readFolders, copyFilesToFolder } = require("../../github");
-const { processYamlTemplate } = require("../../yamlProcessor");
-
-/**
- * Handler for /request-research command
- * Opens the research request modal with pre-filled user information
- */
-const requestResearchHandler = async ({ ack, command, client }) => {
-  await ack();
-
-  try {
-    // Fetch user info to populate the submitted by field
-    const userInfo = await client.users.info({ user: command.user_id });
-    const userName = userInfo.user.real_name || userInfo.user.name;
-    const userTitle = userInfo.user.profile.title || 'Team Member';
-    const displayName = userTitle ? `${userName}, ${userTitle}` : userName;
-
-    // Create a copy of the modal and update the submitted_by field
-    const modalView = JSON.parse(JSON.stringify(requestResearchModal));
-
-    // Find and update the submitted_by input block
-    const submittedByBlock = modalView.blocks.find(block => block.block_id === 'submitted_by_block');
-    if (submittedByBlock) {
-      submittedByBlock.element.initial_value = displayName;
-    }
-
-    await client.views.open({
-      trigger_id: command.trigger_id,
-      view: {
-        ...modalView,
-        private_metadata: JSON.stringify({
-          channelId: command.channel_id,
-          userId: command.user_id,
-          userName: userName,
-          userTitle: userTitle,
-        }),
-      }
-    });
-  } catch (error) {
-    console.error("Error opening request research modal:", error.data || error);
-  }
-};
-
-/**
- * Handler for research request modal submission
- * Processes the request and sends notifications to relevant parties
- */
-const handleRequestResearchSubmission = async ({ ack, body, view, client }) => {
-  await ack();
-
-  const values = view.state.values;
-  const { channelId, userId, userName, userTitle } = JSON.parse(view.private_metadata);
-
-  // Helper function to extract values from different input types
-  const extract = (blockId, actionId) => {
-    const block = values[blockId];
-    if (!block) return '';
-    
-    const action = block[actionId];
-    if (!action) return '';
-    
-    // Handle different input types
-    if (action.value !== undefined) return action.value;
-    if (action.selected_option !== undefined) return action.selected_option.value;
-    if (action.selected_date !== undefined) return action.selected_date;
-    
-    return '';
-  };
-
-  const requestData = {
-    project_title: extract('project_title_block', 'project_title_input'),
-    problem_description: extract('problem_description_block', 'problem_description_input'),
-    affected_users: extract('affected_users_block', 'affected_users_input'),
-    decisions_to_inform: extract('decisions_block', 'decisions_input'),
-    urgency: extract('urgency_block', 'urgency_select'),
-    deadline: extract('deadline_block', 'deadline_picker'),
-    existing_knowledge: extract('existing_knowledge_block', 'existing_knowledge_input'),
-    prepared_by: extract('submitted_by_block', 'submitted_by_input'),
-    requestor_name: extract('submitted_by_block', 'submitted_by_input'),
-    requestedBy: userId,
-    channelId: channelId,
-    // Legacy fields for backward compatibility
-    timelineNeeded: extract('urgency_block', 'urgency_select'), // Map urgency to timelineNeeded
-  };
-
-  console.log('📬 Research Request Submitted:', requestData);
-
-  try {
-    // Get user info
-    const userInfo = await client.users.info({ user: userId });
-    const submitterName = userInfo.user.real_name || userInfo.user.name;
-
-    // Process YAML template to create the research request file
-    let requestUrl = null;
-    let requestPath = null;
-    try {
-      const yamlFile = await fetchFileFromRepo(getConfigRepo(), YAML_TEMPLATE_PATH, "research_request.yaml");
-      
-      // Process YAML template — write to content repo, not config repo
-      const sanitizedTitle = (requestData.project_title || 'research-request')
-        .toLowerCase().replace(/\s+/g, '-').replace(/[^a-z0-9-]/g, '');
-      const renderedYaml = await processYamlTemplate(
-        yamlFile.content,
-        requestData,
-        sanitizedTitle,
-        "research-requests"
-      );
-      
-      requestUrl = renderedYaml.result.url;
-      requestPath = renderedYaml.result.path;
-      
-      console.log('✅ Research request file created:', requestUrl);
-    } catch (yamlError) {
-      console.error('⚠️ Error processing research request YAML:', yamlError);
-      // Continue with notification even if YAML processing fails
-    }
-
-    // Send confirmation to the user
-    await client.chat.postMessage({
-      channel: userId,
-      text: `✅ *Your research request has been submitted!*\n\n*Project:* ${requestData.project_title}\n\nThe research team will review your request and get back to you soon.${requestUrl ? `\n\n📄 <${requestUrl}|View Request on GitHub>` : ''}`
-    });
-
-    // Send notification to research team channel with "Create Brief" button
-    const researchTeamChannelId = process.env.RESEARCH_TEAM_CHANNEL_ID;
-    const isValidChannelId = researchTeamChannelId && researchTeamChannelId !== 'C1234567890';
-
-    if (!isValidChannelId) {
-      console.warn('⚠️ RESEARCH_TEAM_CHANNEL_ID not configured — skipping research team notification. Set this env var to a real Slack channel ID to enable notifications.');
-    }
-
-    const notificationBlocks = [
-      {
-        type: "header",
-        text: {
-          type: "plain_text",
-          text: "📬 New Research Request",
-          emoji: true,
-        },
-      },
-      {
-        type: "section",
-        text: {
-          type: "mrkdwn",
-          text: `*From:* <@${userId}>\n*Project:* ${requestData.project_title}`,
-        },
-      },
-      {
-        type: "section",
-        fields: [
-          {
-            type: "mrkdwn",
-            text: `*Problem/Opportunity:*\n${(requestData.problem_description || '').substring(0, 200)}${(requestData.problem_description || '').length > 200 ? '...' : ''}`,
-          },
-          {
-            type: "mrkdwn",
-            text: `*Urgency:*\n${requestData.urgency ? requestData.urgency.charAt(0).toUpperCase() + requestData.urgency.slice(1) : 'Not specified'}`,
-          },
-        ],
-      },
-      ...(requestData.affected_users ? [{
-        type: "section",
-        text: {
-          type: "mrkdwn",
-          text: `*Affected Users:*\n${requestData.affected_users}`,
-        },
-      }] : []),
-    ];
-
-    // Add GitHub link section if request file was created
-    if (requestUrl) {
-      notificationBlocks.push({
-        type: "section",
-        text: {
-          type: "mrkdwn",
-          text: `📄 <${requestUrl}|View Full Research Request on GitHub>`,
-        },
-      });
-    }
-
-    notificationBlocks.push(
-      {
-        type: "divider",
-      },
-      {
-        type: "actions",
-        elements: [
-          {
-            type: "button",
-            text: {
-              type: "plain_text",
-              text: "📄 Create Brief from Request",
-              emoji: true,
-            },
-            style: "primary",
-            action_id: "create_brief_from_request",
-            value: JSON.stringify({
-              ...requestData,
-              requestUrl, // Include the request file URL
-              requestPath, // Include the request file path
-            }),
-          }
-        ],
-      }
-    );
-
-    if (isValidChannelId) {
-      try {
-        await client.chat.postMessage({
-          channel: researchTeamChannelId,
-          text: `📬 New Research Request from ${submitterName}`,
-          blocks: notificationBlocks,
-        });
-        console.log(`✅ Research request notification sent to channel ${researchTeamChannelId}`);
-      } catch (channelError) {
-        console.error(`❌ Failed to send research team notification to ${researchTeamChannelId}:`, channelError.message);
-        // Don't fail the whole request — the submitter already got their confirmation
-      }
-    }
-
-  } catch (error) {
-    console.error('Error processing research request:', error);
-    await client.chat.postMessage({
-      channel: userId,
-      text: `❌ There was an error submitting your research request. Please try again or contact support.`
-    });
-  }
-};
-
-/**
- * Handler for "Create Brief from Request" button click
- * Opens the research brief modal with pre-filled data from the request
- */
-const handleCreateBriefFromRequest = async ({ ack, body, client }) => {
-  await ack();
-
-  try {
-    const requestData = JSON.parse(body.actions[0].value);
-    const channelId = body.channel.id;
-    const userId = body.user.id;
-
-    // Create a copy of the research brief modal
-    const modalView = JSON.parse(JSON.stringify(researchBriefModal));
-
-    // Filter blocks to remove request link and stakeholder if not available
-    const filteredBlocks = modalView.blocks.filter(block => {
-      // Remove request link block if not available
-      if (block.block_id === 'request_link_block' && !requestData.requestUrl) {
-        return false;
-      }
-      // Remove stakeholder block if not available
-      if (block.block_id === 'stakeholder_block' && !requestData.prepared_by) {
-        return false;
-      }
-      return true;
-    });
-    
-    // Update remaining blocks
-    filteredBlocks.forEach(block => {
-      if (block.type === 'input' && block.block_id) {
-        // Pre-fill study title
-        if (block.block_id === 'study_title_block' && block.element?.action_id === 'study_title_input') {
-          block.element.initial_value = requestData.project_title || '';
-          block.hint = {
-            type: "plain_text",
-            text: "Auto-filled from research request",
-          };
-        }
-        // Pre-fill stakeholder (only if block exists)
-        else if (block.block_id === 'stakeholder_block' && block.element?.action_id === 'stakeholder_input') {
-          block.element.initial_value = requestData.prepared_by || '';
-        }
-        // Clear all other fields
-        else {
-          // Remove initial_value/initial_date for all other fields
-          if (block.element) {
-            delete block.element.initial_value;
-            delete block.element.initial_date;
-          }
-        }
-      }
-      // Handle request link as clickable section (only if block exists)
-      else if (block.type === 'section' && block.block_id === 'request_link_block' && requestData.requestUrl) {
-        block.text.text = `<${requestData.requestUrl}|:page_facing_up: View original research request on GitHub>`;
-      }
-    });
-    
-    modalView.blocks = filteredBlocks;
-
-    console.log('📄 Opening research brief modal with pre-filled data from request');
-
-    // Open the research brief modal with pre-filled data
-    await client.views.open({
-      trigger_id: body.trigger_id,
-      view: {
-        ...modalView,
-        private_metadata: JSON.stringify({
-          channelId,
-          userId,
-          isFromRequest: true,
-          requestData,
-          // Use a default path for requests (no study exists yet)
-          studyName: requestData.project_title || 'Research Request',
-        }),
-      },
-    });
-  } catch (error) {
-    console.error("Error opening research brief modal from request:", error.data || error);
-  }
-};
-
-/**
- * Handler for "Create Study from Request" button click
- * Opens the create study modal with pre-filled data from the request
- */
-const handleCreateStudyFromRequest = async ({ ack, body, client }) => {
-  await ack();
-
-  try {
-    const requestData = JSON.parse(body.actions[0].value);
-    const channelId = body.channel.id;
-    const userId = body.user.id;
-
-    // Normalize field names for modal compatibility
-    if (requestData.prepared_by && !requestData.submittedBy) {
-      requestData.submittedBy = requestData.prepared_by;
-    }
-    if (requestData.project_title && !requestData.projectTitle) {
-      requestData.projectTitle = requestData.project_title;
-    }
-
-    console.log('📋 Opening create study modal with data:', {
-      prepared_by: requestData.prepared_by,
-      submittedBy: requestData.submittedBy,
-      projectTitle: requestData.projectTitle,
-      project_title: requestData.project_title,
-    });
-
-    // Open the create study from request modal with pre-filled data
-    await client.views.open({
-      trigger_id: body.trigger_id,
-      view: {
-        ...createStudyFromRequestModal(requestData),
-        private_metadata: JSON.stringify({
-          channelId,
-          userId,
-          isFromRequest: true,
-          requestData,
-        }),
-      },
-    });
-  } catch (error) {
-    console.error("Error opening create study from request modal:", error.data || error);
-  }
-};
-
-
-module.exports = {
-  requestResearchHandler,
-  handleRequestResearchSubmission,
-  handleCreateBriefFromRequest,
-  handleCreateStudyFromRequest,
-};
-
